+++ CHANGELOG.md
@@ -37,7 +37,7 @@ Previous classification is not required if changes are simple or all belong to t
 - Fixed some typos and grammatical errors (mostly on code comments).
 - Added new extension method `AddDefaultDocumentConnectorProvider` in `Encamina.Enmarcha.SemanticKernel.Connectors.Document` to get access to a default implementation of a `IDocumentConnector`.
 - Updated sample projects with latest changes.
-- Added optional parameter to `AddDefaultDocumentConnectorProvider` to pass it a function to calculate the length of a text and inject it as a dependency.
+- Added optional parameter to `AddDefaultDocumentConnectorProvider` and `AddDefaultDocumentContentExtractor` to pass it a function to calculate the length of a text and inject it as a dependency.
 
 ## [8.1.1]